+++ lib/src/model/category.g.dart
@@ -8,14 +8,14 @@ part of 'category.dart';
 
 CategoriesResponse _$CategoriesResponseFromJson(Map<String, dynamic> json) =>
     CategoriesResponse(
-      categories: (json['categories'] as List<dynamic>)
+      items: (json['categories'] as List<dynamic>)
           .map((e) => Category.fromJson(e as Map<String, dynamic>))
           .toList(),
     );
 
 Map<String, dynamic> _$CategoriesResponseToJson(CategoriesResponse instance) =>
     <String, dynamic>{
-      'categories': instance.categories,
+      'categories': instance.items,
     };
 
 Category _$CategoryFromJson(Map<String, dynamic> json) => Category(